Set up a basic tool for printing iset downsets.
It's called "functor" because it documents the functor from iset.mm to
jbobau.mm. The idea is that, by visiting the downset, we can print out
all of the theorems and axioms which are required in order to implement
anything from iset.mm in terms of Lojban. This is a formalization and
mechanization of an existing process already acting on jbobau.mm.

Fixes #6.