Implement ~nsyl.
This was the first time that I used the automated chain-finding tool
introduced a few commits ago, and it performed wonderfully. The only
limitation is that I found myself loading up iset.mm in order to

    show proof nsyl

Or other dependencies, because it wasn't always obvious how everything
fits together. Fixing this would require augmenting the tool to show
proof trees, which is not what I want to do at this time of night.

Fixes #5. All done!