Start porting the complex-number axioms.
Lemme say a few words here. So, there's three different ZF-style
approaches to this so far in Metamath, and they all bite because they
aren't neutral with respect to the internal structure of sets. They make
a large deal out of the nature of N < Z < Q < R < C, which is certainly
gorgeous, but completely miss:

    SemiRing < Ring < Field < Karoubi(Field) < Gauss(Field)

*That* is what I actually care about displaying, not the mere inclusion
of one historically-important number system into another.