Add and document double-negation translation.
There's a mountain of lore on this, but it really just boils down to a
nifty simple little (endo)functor which takes any proven statement and
uses its proof to refute any claimed counterexamples. In classical
logic, that's sufficient to show that something is proven, and so the
functor connects classical and intuitionistic logic.