Improve ~bnk syntax hint.
~bnk doesn't normally provide the desired syntax hint for {naku} because
it carried an extra floating hypothesis ~bnk.0. By removing that
hypothesis and unscoping ~bnk, we no longer have to explicitly assign it
while writing proofs. The downside is that, one time only, I had to fix
up a dozen proofs. However, now it's done, and I'm done for the day.