Sketch definition of {cajni}.
I think that it might be worth double-checking exactly what the kurtyvla
offered compared to what I've chosen as definitions. I recall that I was
trying to offer words that highlighted desirable algebras, but I'm not
sure if I'm ready to define groups as {dukni}, which is the endgame.

I did try to define inverses, but I need to highlight the identity
element in the same scope or else it's possible that identities/inverses
aren't unique. This is something of an opportunity since it means that
we'll have a chance to non-trivially show that every dukni1 is also a
sezni1 over the same domain.