Add ~subt.
We're going to add Margaris to the bibliography. I haven't actually read
the book, but it's the best citation for some of these side theorems
which aren't simple lemmas nor restatements of standard properties.

I'm starting to have a cycle for incrementally adding new theorems. It
is not trivial but it has greatly reduced the mental load of operating
the proof assistant, as well as reduced the amount of planning time
required to prepare the proof.